Test Commit

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/common/FlickerServo.java
@@ -18,37 +18,41 @@ public class FlickerServo {
     }
 
     public State flickerState;
-    private int maxFlicks;
+    public int maxFlicks;
     private int curFlicks;
     private ElapsedTime timer;
     private Telemetry telemetry;
     private Servo flicker;
+    private boolean flicking;
+    private boolean nodrop;
 
     //degrees for going foward starts at 1.0
-    private double flickPosition = 0.75;
-    private double basePosition = 1.0;
+    private double flickPosition = 0.65; //0.75;
+    private double basePosition = 1.0; //1.0;
+    private double startPosition = 0.75;
+
     //time between each flick
-    private double fwdTime = 120;
-    private double backTime = 135;
-    private double pauseTime = 50;
+    private double fwdTime = 100; // 120
+    private double backTime = 450; // 135
     private Elevator elevator;
 
     public FlickerServo(HardwareMap hardwareMap, Elevator elev, Telemetry tele)
     {
         flicker = hardwareMap.get(Servo.class, "flicker");
         this.telemetry = tele;
-        this.flicker.setPosition(this.basePosition);
+        this.flicker.setPosition(this.startPosition);
         this.flickerState = State.STATE_FLICKER_STOPPED;
         this.maxFlicks = 0;
         this.timer = new ElapsedTime(ElapsedTime.Resolution.MILLISECONDS);
         this.elevator = elev;
+        this.flicking = false;
+        this.nodrop = false;
     }
 
     public void flick(int times)
     {
         if (this.flickerState != State.STATE_FLICKER_STOPPED)
         {
-            this.maxFlicks += times;
             return;
         }
 
@@ -59,6 +63,18 @@ public class FlickerServo {
         /* Lift the elevator */
         this.elevator.lift();
         this.flickerState = State.STATE_FLICKER_ELEVATOR;
+        this.flicking = true;
+    }
+
+    public void forceFlick(int times)
+    {
+        /* Set maxflicks and initialize current flicks to 0 */
+        this.maxFlicks = times;
+        this.curFlicks = 0;
+        this.flicking = true;
+        this.flickerState = State.STATE_FLICKER_BACK;
+        this.flicker.setPosition(this.basePosition);
+        this.timer.reset();
     }
 
     public void stop()
@@ -70,11 +86,17 @@ public class FlickerServo {
         this.curFlicks = 0;
         this.flickerState = State.STATE_FLICKER_BACK;
         this.timer.reset();
+        this.flicking = false;
     }
 
     public boolean isStopped()
     {
         return flickerState == State.STATE_FLICKER_STOPPED;
+    }
+
+    public void nodrop()
+    {
+        nodrop = true;
     }
 
     public void update()
@@ -83,8 +105,8 @@ public class FlickerServo {
             case STATE_FLICKER_ELEVATOR:
                 if (this.elevator.eleState == Elevator.State.STATE_ELEVATOR_SWIVELED) {
                     /* Set flicker state to FWD and set position */
-                    this.flickerState = State.STATE_FLICKER_FWD;
-                    this.flicker.setPosition(this.flickPosition);
+                    this.flickerState = State.STATE_FLICKER_BACK;
+                    this.flicker.setPosition(this.basePosition);
                     this.timer.reset();
                 }
             break;
@@ -92,24 +114,11 @@ public class FlickerServo {
             case STATE_FLICKER_FWD:
                 if (this.timer.milliseconds() >= this.fwdTime)
                 {
-                    /* Increment current flick count */
-                    this.curFlicks = this.curFlicks + 1;
-
-                    /* Move flicker back */
-                    this.flickerState = State.STATE_FLICKER_BACK;
-                    this.flicker.setPosition(this.basePosition);
-                    this.timer.reset();
-                }
-            break;
-
-            case STATE_FLICKER_BACK:
-                if (this.timer.milliseconds() >= this.backTime)
-                {
                     if (this.curFlicks != this.maxFlicks)
                     {
                         /* Still have flicks to do - start going forward */
-                        this.flicker.setPosition(this.flickPosition);
-                        this.flickerState = State.STATE_FLICKER_FWD;
+                        this.flicker.setPosition(this.basePosition);
+                        this.flickerState = State.STATE_FLICKER_BACK;
                         this.timer.reset();
                     }
                     else
@@ -117,15 +126,33 @@ public class FlickerServo {
                         this.maxFlicks = 0;
                         this.curFlicks = 0;
                         this.flickerState = State.STATE_FLICKER_STOPPED;
+                        this.flicker.setPosition(this.startPosition);
                     }
+                }
+            break;
+
+            case STATE_FLICKER_BACK:
+                if (this.timer.milliseconds() >= this.backTime)
+                {
+                    /* Increment current flick count */
+                    this.curFlicks = this.curFlicks + 1;
+
+                    /* Move flicker back */
+                    this.flickerState = State.STATE_FLICKER_FWD;
+                    this.flicker.setPosition(this.flickPosition);
+                    this.timer.reset();
                 }
             break;
 
             case STATE_FLICKER_STOPPED:
                 /* Flicker is stopped - nothing to do. Move the elevator down */
-                if (this.elevator.eleState == Elevator.State.STATE_ELEVATOR_SWIVELED)
+                if ((this.elevator.eleState == Elevator.State.STATE_ELEVATOR_SWIVELED) && flicking)
                 {
-                    this.elevator.drop();
+                    if (!nodrop) {
+                        this.elevator.drop();
+                        this.flicking = false;
+                    }
+                    nodrop = true;
                 }
             break;
         }
